review comments applied

+++ src/hotspot/share/gc/shared/gc_globals.hpp
@@ -682,10 +682,10 @@
   develop(uintx, GCExpandToAllocateDelayMillis, 0,                          \
           "Delay between expansion and allocation (in milliseconds)")       \
                                                                             \
-  product(uint, GCDrainStackTargetSize, 64,                                 \
+  product(size_t, GCDrainStackTargetSize, 64,                               \
           "Number of entries we will try to leave on the stack "            \
           "during parallel gc")                                             \
-          range(0, UINT_MAX)                                                \
+          range(0, SIZE_MAX)                                                \
                                                                             \
   product(uint, GCCardSizeInBytes, 512,                                     \
           "Card table entry size (in bytes) for card based collectors")     \
